Fix various warnings found during a release build (mostly CGRect/NSRect and properties)

+++ Source/Core/NBEngineHighlighter.m
@@ -59,7 +59,7 @@
     return self;
 }
 
-- (unsigned long)retainCount
+- (NSUInteger)retainCount
 {
     return ULONG_MAX;
 }
